feat(cli): Support trace file URLs
