+++ src/utils/browserFetcher.ts
@@ -19,8 +19,7 @@ import extract from 'extract-zip';
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
-import ProgressBar from 'progress';
-import { downloadFile, existsAsync } from './utils';
+import { downloadFile, existsAsync, getDownloadProgress, getErrorMessage } from './utils';
 import { debugLogger } from './debugLogger';
 
 export async function downloadBrowserWithProgressBar(title: string, browserDirectory: string, executablePath: string, downloadURL: string, downloadFileName: string): Promise<boolean> {
@@ -31,36 +30,17 @@ export async function downloadBrowserWithProgressBar(title: string, browserDirec
     return false;
   }
 
-  let progressBar: ProgressBar;
-  let lastDownloadedBytes = 0;
-
-  function progress(downloadedBytes: number, totalBytes: number) {
-    if (!process.stderr.isTTY)
-      return;
-    if (!progressBar) {
-      progressBar = new ProgressBar(`Downloading ${progressBarName} - ${toMegabytes(totalBytes)} [:bar] :percent :etas `, {
-        complete: '=',
-        incomplete: ' ',
-        width: 20,
-        total: totalBytes,
-      });
-    }
-    const delta = downloadedBytes - lastDownloadedBytes;
-    lastDownloadedBytes = downloadedBytes;
-    progressBar.tick(delta);
-  }
-
   const url = downloadURL;
   const zipPath = path.join(os.tmpdir(), downloadFileName);
   try {
     for (let attempt = 1, N = 3; attempt <= N; ++attempt) {
       debugLogger.log('install', `downloading ${progressBarName} - attempt #${attempt}`);
-      const {error} = await downloadFile(url, zipPath, {progressCallback: progress, log: debugLogger.log.bind(debugLogger, 'install')});
+      const {error} = await downloadFile(url, zipPath, {progressCallback: getDownloadProgress(progressBarName), log: debugLogger.log.bind(debugLogger, 'install')});
       if (!error) {
         debugLogger.log('install', `SUCCESS downloading ${progressBarName}`);
         break;
       }
-      const errorMessage = typeof error === 'object' && typeof error.message === 'string' ? error.message : '';
+      const errorMessage = getErrorMessage(error);
       debugLogger.log('install', `attempt #${attempt} - ERROR: ${errorMessage}`);
       if (attempt < N && (errorMessage.includes('ECONNRESET') || errorMessage.includes('ETIMEDOUT'))) {
         // Maximum delay is 3rd retry: 1337.5ms
@@ -89,10 +69,6 @@ export async function downloadBrowserWithProgressBar(title: string, browserDirec
   return true;
 }
 
-function toMegabytes(bytes: number) {
-  const mb = bytes / 1024 / 1024;
-  return `${Math.round(mb * 10) / 10} Mb`;
-}
 
 export function logPolitely(toBeLogged: string) {
   const logLevel = process.env.npm_config_loglevel;
